format

+++ Interfaces/ICommunityIssuance.sol
@@ -2,19 +2,18 @@
 
 pragma solidity 0.6.11;
 
-interface ICommunityIssuance { 
-    
+interface ICommunityIssuance {
     // --- Events ---
-    
+
     event LQTYTokenAddressSet(address _lqtyTokenAddress);
     event StabilityPoolAddressSet(address _stabilityPoolAddress);
-    event TotalLQTYIssuedUpdated(uint _totalLQTYIssued);
+    event TotalLQTYIssuedUpdated(uint256 _totalLQTYIssued);
 
     // --- Functions ---
 
     function setAddresses(address _lqtyTokenAddress, address _stabilityPoolAddress) external;
 
-    function issueLQTY() external returns (uint);
+    function issueLQTY() external returns (uint256);
 
-    function sendLQTY(address _account, uint _LQTYamount) external;
+    function sendLQTY(address _account, uint256 _LQTYamount) external;
 }
