+++ Interfaces/IDefaultPool.sol
@@ -4,13 +4,12 @@ pragma solidity 0.6.11;
 
 import "./IPool.sol";
 
-
 interface IDefaultPool is IPool {
     // --- Events ---
     event TroveManagerAddressChanged(address _newTroveManagerAddress);
-    event DefaultPoolLUSDDebtUpdated(uint _LUSDDebt);
-    event DefaultPoolETHBalanceUpdated(uint _ETH);
+    event DefaultPoolLUSDDebtUpdated(uint256 _LUSDDebt);
+    event DefaultPoolETHBalanceUpdated(uint256 _ETH);
 
     // --- Functions ---
-    function sendETHToActivePool(uint _amount) external;
+    function sendETHToActivePool(uint256 _amount) external;
 }
