+++ Proxy/BorrowerOperationsScript.sol
@@ -5,7 +5,6 @@ pragma solidity 0.6.11;
 import "../Dependencies/CheckContract.sol";
 import "../Interfaces/IBorrowerOperations.sol";
 
-
 contract BorrowerOperationsScript is CheckContract {
     IBorrowerOperations immutable borrowerOperations;
 
@@ -14,23 +13,23 @@ contract BorrowerOperationsScript is CheckContract {
         borrowerOperations = _borrowerOperations;
     }
 
-    function openTrove(uint _maxFee, uint _LUSDAmount, address _upperHint, address _lowerHint) external payable {
-        borrowerOperations.openTrove{ value: msg.value }(_maxFee, _LUSDAmount, _upperHint, _lowerHint);
+    function openTrove(uint256 _maxFee, uint256 _LUSDAmount, address _upperHint, address _lowerHint) external payable {
+        borrowerOperations.openTrove{value: msg.value}(_maxFee, _LUSDAmount, _upperHint, _lowerHint);
     }
 
     function addColl(address _upperHint, address _lowerHint) external payable {
-        borrowerOperations.addColl{ value: msg.value }(_upperHint, _lowerHint);
+        borrowerOperations.addColl{value: msg.value}(_upperHint, _lowerHint);
     }
 
-    function withdrawColl(uint _amount, address _upperHint, address _lowerHint) external {
+    function withdrawColl(uint256 _amount, address _upperHint, address _lowerHint) external {
         borrowerOperations.withdrawColl(_amount, _upperHint, _lowerHint);
     }
 
-    function withdrawLUSD(uint _maxFee, uint _amount, address _upperHint, address _lowerHint) external {
+    function withdrawLUSD(uint256 _maxFee, uint256 _amount, address _upperHint, address _lowerHint) external {
         borrowerOperations.withdrawLUSD(_maxFee, _amount, _upperHint, _lowerHint);
     }
 
-    function repayLUSD(uint _amount, address _upperHint, address _lowerHint) external {
+    function repayLUSD(uint256 _amount, address _upperHint, address _lowerHint) external {
         borrowerOperations.repayLUSD(_amount, _upperHint, _lowerHint);
     }
 
@@ -38,8 +37,17 @@ contract BorrowerOperationsScript is CheckContract {
         borrowerOperations.closeTrove();
     }
 
-    function adjustTrove(uint _maxFee, uint _collWithdrawal, uint _debtChange, bool isDebtIncrease, address _upperHint, address _lowerHint) external payable {
-        borrowerOperations.adjustTrove{ value: msg.value }(_maxFee, _collWithdrawal, _debtChange, isDebtIncrease, _upperHint, _lowerHint);
+    function adjustTrove(
+        uint256 _maxFee,
+        uint256 _collWithdrawal,
+        uint256 _debtChange,
+        bool isDebtIncrease,
+        address _upperHint,
+        address _lowerHint
+    ) external payable {
+        borrowerOperations.adjustTrove{value: msg.value}(
+            _maxFee, _collWithdrawal, _debtChange, isDebtIncrease, _upperHint, _lowerHint
+        );
     }
 
     function claimCollateral() external {
