+++ Proxy/LQTYStakingScript.sol
@@ -5,7 +5,6 @@ pragma solidity 0.6.11;
 import "../Dependencies/CheckContract.sol";
 import "../Interfaces/ILQTYStaking.sol";
 
-
 contract LQTYStakingScript is CheckContract {
     ILQTYStaking immutable LQTYStaking;
 
@@ -14,7 +13,7 @@ contract LQTYStakingScript is CheckContract {
         LQTYStaking = ILQTYStaking(_lqtyStakingAddress);
     }
 
-    function stake(uint _LQTYamount) external {
+    function stake(uint256 _LQTYamount) external {
         LQTYStaking.stake(_LQTYamount);
     }
 }
