+++ TestContracts/ActivePoolTester.sol
@@ -5,9 +5,8 @@ pragma solidity 0.6.11;
 import "../ActivePool.sol";
 
 contract ActivePoolTester is ActivePool {
-    
-    function unprotectedIncreaseLUSDDebt(uint _amount) external {
-        LUSDDebt  = LUSDDebt.add(_amount);
+    function unprotectedIncreaseLUSDDebt(uint256 _amount) external {
+        LUSDDebt = LUSDDebt.add(_amount);
     }
 
     function unprotectedPayable() external payable {
