+++ TestContracts/BorrowerOperationsTester.sol
@@ -7,57 +7,43 @@ import "../BorrowerOperations.sol";
 /* Tester contract inherits from BorrowerOperations, and provides external functions 
 for testing the parent's internal functions. */
 contract BorrowerOperationsTester is BorrowerOperations {
-
-    function getNewICRFromTroveChange
-    (
-        uint _coll, 
-        uint _debt, 
-        uint _collChange, 
-        bool isCollIncrease, 
-        uint _debtChange, 
-        bool isDebtIncrease, 
-        uint _price
-    ) 
-    external
-    pure
-    returns (uint)
-    {
+    function getNewICRFromTroveChange(
+        uint256 _coll,
+        uint256 _debt,
+        uint256 _collChange,
+        bool isCollIncrease,
+        uint256 _debtChange,
+        bool isDebtIncrease,
+        uint256 _price
+    ) external pure returns (uint256) {
         return _getNewICRFromTroveChange(_coll, _debt, _collChange, isCollIncrease, _debtChange, isDebtIncrease, _price);
     }
 
-    function getNewTCRFromTroveChange
-    (
-        uint _collChange, 
-        bool isCollIncrease,  
-        uint _debtChange, 
-        bool isDebtIncrease, 
-        uint _price
-    ) 
-    external 
-    view
-    returns (uint) 
-    {
+    function getNewTCRFromTroveChange(
+        uint256 _collChange,
+        bool isCollIncrease,
+        uint256 _debtChange,
+        bool isDebtIncrease,
+        uint256 _price
+    ) external view returns (uint256) {
         return _getNewTCRFromTroveChange(_collChange, isCollIncrease, _debtChange, isDebtIncrease, _price);
     }
 
-    function getUSDValue(uint _coll, uint _price) external pure returns (uint) {
+    function getUSDValue(uint256 _coll, uint256 _price) external pure returns (uint256) {
         return _getUSDValue(_coll, _price);
     }
 
-    function callInternalAdjustLoan
-    (
-        address _borrower, 
-        uint _collWithdrawal, 
-        uint _debtChange, 
-        bool _isDebtIncrease, 
+    function callInternalAdjustLoan(
+        address _borrower,
+        uint256 _collWithdrawal,
+        uint256 _debtChange,
+        bool _isDebtIncrease,
         address _upperHint,
-        address _lowerHint)
-        external 
-    {
+        address _lowerHint
+    ) external {
         _adjustTrove(_borrower, _collWithdrawal, _debtChange, _isDebtIncrease, _upperHint, _lowerHint, 0);
     }
 
-
     // Payable fallback function
-    receive() external payable { }
+    receive() external payable {}
 }
