+++ TestContracts/DefaultPoolTester.sol
@@ -5,9 +5,8 @@ pragma solidity 0.6.11;
 import "../DefaultPool.sol";
 
 contract DefaultPoolTester is DefaultPool {
-    
-    function unprotectedIncreaseLUSDDebt(uint _amount) external {
-        LUSDDebt  = LUSDDebt.add(_amount);
+    function unprotectedIncreaseLUSDDebt(uint256 _amount) external {
+        LUSDDebt = LUSDDebt.add(_amount);
     }
 
     function unprotectedPayable() external payable {
